Make the test suite not fail on ubuntu
libdl is a fairly common library; if you forget to explicitly link to
it, chances are high that some other library in your dependency chain
will need it anyway, so it will be in your process space and all is
well. This is the case on most Linux distributions.

However, on ubuntu it happens to not be the case for us, which means the
test suite fails to link, let alone run.

Fix by explicitly telling the linker that libbeidcommon.so uses libdl,
as we should.