Counting function calls and assertions sucks
Previously, verbose_assert() and ckrv_decode() would output "assertion
X" or "function X", and nothing more. This works, but is annoying in
case a call fails and we don't exactly know for sure where.

Make it more verbose, by printing the assertion and/or function call to
stdout instead. This required a refactoring of the
ckrv_decode()/check_rv stuff, so it would actually get the function
name -- and be slightly prettier, too.