Fix warnings
Modifying a string literal from code == segfault. Modern compilers will
warn that pointing a char* (as opposed to a const char*) to a string
literal is deprecated, for that reason.

To avoid the warning (and, occasionally, the bug), mark them as const
char* rather than char.

This should also make Coverity somewhat more happy :-)