Add Travis configuration
We already have a buildbot running on-premise, which builds packages for
a multitude of systems; but since it is behind a firewall, it does not
(and cannot) integrate with github.

Add travis instead, which can't test as much as the buildbot, but will
give us at least a useful basic "it compiles!" test.