Reorder things slightly
The "slotevent" test would eject the card, not wait for the result, and
then exit. This would cause the "mechlist" test to fail due to a
CKR_DEVICE_ERROR.

We should possibly figure out if we can actually fix the
CKR_DEVICE_ERROR (although that may be something we can't actually do),
but for now at least keep things in a proper state when we SKIP a test.