Check a number of incorrect initialization calls
Added after look at code coverage, and implementing a test for code not
currently covered.

Also adds a bug fix in the initialization: if initialization fails
because of incorrect arguments, any further calls to C_Initialize in the
same process space would fail because our internal state would be set to
BEIDP11_INITIALIZING and not be reset. Fix by resetting it (after the
"cleanup:" label) to BEIDP11_NOT_INITIALIZED if our return value is not
CKR_OK.