Switch to gtk+ 3.0 where possible
We put in some work a while back to make it possible to compile the
middleware against gtk+ 3.0 rather than 2.0. However, since Ubuntu 12.04
and RHEL/CentOS 6 do not yet have that version of GTK, we cannot drop
gtk+ 2.0 support yet.

Since stuff currently should just compile against gtk+ 3.0, switch the
configure script and the packaging dependencies so that they try to use
gtk+ 3.0 if available, but will fall back to gtk+ 2.0 if it isn't
available.