Do conditional gtk2 differently
RPM doesn't know alternative build-dependencies, like the .deb format
does (either that, or I didn't look hard enough). However, it _is_
possible to differentiate based on distribution by using some macros in
the specfile.

Do that, so that we use gtk2 on el6 only, rather than trying to use the
"if gtk2 is old enough, use that" heuristic.