Split off the label categorisation
In the new eID viewer, we want to issue different calls based on whether
something is a string or not, too.

Rather than reimplementing this categorisation (which we'd already done
for the test suite), split it into a separate file which we can then
reuse for the viewer.

Still TODO: update Windows and OSX build files for this as well.