Use a g_object_set() variant that is thread-safe
GTK+ doesn't like its properties being set from anywhere but its own
main loop. Fortunately, there is a function call that allows you to
marshal a function in another thread.

Use that function to write a g_object_set_threaded() which will call the
method on the main thread (or, at the very least, in the mainloop
context).