Fix implementation of threaded g_object_set()
va_lists aren't supposed to work across threading boundaries, so don't
try that. We (usually) need to set only one object, anyway.

Additionally, we may need to free a value, and we can't do that, since
hey, this might be in a different thread. Instruct the function to do so
for us instead.