Improve namespacing
It just so happened that the PKCS#11 module had a function
"p11_close_session()", which was incorrectly exported and of which we
reused the name. As such, that function was called with garbage as its
first parameter, which caused a crash.

Prefix all p11 functions with "eid_vwr" so this doesn't happen again.

TODO: Make sure the PKCS#11 doesn't export such non-PKCS#11 functions
anymore.