Defend against uninitialized calls
We shouldn't blatantly assume that all these functions are always
available, otherwise we're not forward compatible.

Every time we call the function, we should first check whether the
callbacks have been set, and if so, whether the set callbacks actually
contain the requested function. If they don't, don't try to call the
relevant function.