Push statusbar updates to UI code, and make it thread-safe
We previously did status bar updates by way of a "newstatus" callback,
which had a string as argument (to be shown in the UI). This was a bad
idea, for several reasons:

- It makes the displayed string untranslateable
- It makes it difficult to decide whether or not to show a spinner
- It doesn't consider things like "show the correct icon"

Instead, send messages containing the current state for the state
machine to the UI layer, and let the UI layer decide what to do with
those state messages.

This required that "enum states" be brought into the oslayer.h header,
which in turn required that it be namespaced.