Fix location of strdup calls
Calling std::string::c_str() returns a pointer to data that is only
valid for the lifetime of the std::string; once its destructor is
called, the pointer points to junk.

While we did call strdup on the result, we call it in the wrong stack
frame, which means the call may point to memory which has already been
reused. Fix by calling it in the same stack frame where we call c_str()
instead.

As an added benefit, we also fix the dormant bug where we apparently
forgot the strdup() plus the location where we did it twice (and
introduced a memory leak).

CID#89253