Fix C_GetSlotList() call
If we call C_GetSlotList with a NULL pointer as argument, it will
function in a different manner than if we don't do that. Specifically,
it won't return CKR_BUFFER_TOO_SMALL in that case.

Arguably a bug in the PKCS#11 spec, it's easier if we just never call it
with a NULL pointer then.