Improve translation method
Previously, we tried doing the translation thing by manually listing, in
code, the labels that needed to be translated. While this works, it is
duplication of effort (and therefore not ideal).

Rework by parsing the XML string instead (which we already have) and
doing g_object_set() calls for each and every UI element that has the
"translatable" attribute set. This way, to be translated strings are
specified in XML only.