Start over
We wanted to do this in swift first, but swift requires OSX 10.9 or
above. We still support older versions than that at this point, so
that's not an option for now.

Unfortunately, it seems fairly difficult to switch from a swift-based
project to an Objective-C one in Xcode; at least I didn't find out how
to do it. Rather than searching for days on end, just start over (there
wasn't that much code yet anyway).