Cache views
We currently use an O(n^2) algorithm to match views to labels, which is
not ideal. Make it slightly less problematic by at least caching the
results.

To properly fix this, we should probably go over the list of items once
and store them in a dictionary instead, but that's for later.