Generate the XML at TOKEN_WAIT and FILE entry
Previously, we would generate the XML to a filename when the user
requested "save file". This makes it impossible to reuse the XML
generation for other purposes (e.g., drag-and-drop).

Instead, use an xmlTextWriterMemory to do the generation when we enter
the TOKEN_WAIT and FILE states, and store the result in the cache. Then,
when the user requests "save file" operation, we request the cached xml
data.