Fix paths and permissions
The .deb -dev packages already contained the rsaref220 headers, but in a
different location. In order for things not to break (if people were
already using this), retain that path, updating the Makefiles and
pkg-config file.

Fix permissions on newly created files by previous commit, which for
some reason ended up being +x

Update packaging configurations to pick up .pc file and (where this did not yet
happen) headers