Do not validate while previewing
On OSX, accessing our own bundle when we're a plugin is complicated. That means it's difficult to figure out what the XSD location is.

We're only looking for a photo, anyway. If we don't find it, no harm done. So stop checking the XML validity when previewing. We still do check it when opening the file, however.