Split up the "sign" test
We cannot do signature tests if we don't have a PIN code, and we cannot
do state tests if we don't have a robot. However, the two are unrelated,
so we cannot really return a meaningful result if we can do one but not
the other.

Currently, we don't even try to do the state-related tests if we do have
a robot but not a PIN code, which is silly.

As such, split up the test so we can still check the state manipulations
even if we don't have a PIN code.

This will make the test suite fail, but not due to a bug in the test
suite...