Fix verify_null function signature
The verbose_assert() macro, when the TEST_NO_ABORT macro is set, wants to return a value to indicate failure. However, verify_null() was a function that did not have a return value.

Rename it to verify_null_func(), and create a macro that does a check and returns something if necessary to the main function, so we can transparently modify this without having to modify all test suites.