Fix test suite
So there was a bug in the test suite after all: the "privatekey" object
handle that we were passing on to C_SignInit() was from a previous
session, which therefore is invalid.

There does happen to be another bug in the PKCS#11 library in that the
error code returned is probably not valid for the given error condition,
but that's another fix.