Verify that an object is initialized
p11_get_slot_object() looks for an object by a given handle.

It used to be the case that our key objects were initialized when a
session was started, but this is no longer the case; it is now only
initialized when a find operation is performed.

As such, an object handle is only valid when a search has been performed
for that object.

Of course this would have been an error before the change of
initialization, except that the PKCS#11 library just happened to not
break in that case.