Remove opportunistic loading of objects in C_SignInit() call
When we're going to sign, we already expect a private key as an
argument, which implies that C_FindObjects must have run and thus that
the objects have been loaded.

Loading it here does not help, and in fact will result in a user with a
private key found using a C_FindObjects() call issued before an
intervening C_CloseSession() pointing to what should have been an
invalid object.