Fix off-by-one
We check if our current index is smaller than or equal to count, and
then increase the index variable. The next line then writes to (index
variable - 1), which could mean it will write to an element at the index
"count", which is supposed to contain the available number of elements
in the array, not the maximum element index.