Update to support Xcode 7
Since libxml2 and openssl are no longer part of the system libraries
now, this also includes a change to download and build our own copy
rather than trying to use the system-installed ones.

We'll move to using the OSX-native XML and crypto APIs at some undefined
point in the future.