Fix dependencies
tar will unpack a file so that it has the same modification time stamp
as the file had before the tarball was created. This messes with our
dependencies, to the extent where we kept rebuilding everything *all*
the time. This is not useful.

Fiddle with dependencies and stamp files until this doesn't happen
anymore.