Make autoresize of printout work
Thanks to <http://stackoverflow.com/32582043>, we now know how to tell the UI that it needs to resize the print view so it fits just on the page. Whee.

Still TODO: updates to the XIB to match the code, and to make word wrap work properly. That's for the next commit.