Fix ordering
C99 makes implicitly defining a function illegal. Previously, a function
would implicitly be assumed to be declared as a function with the given
parameters returning int. As such, this may happen to compile on
C89-compliant compilers, but it's not good style.

Fix by reordering things so they're declared before use.