Move the flags field to its right position, and use the CDF flag in cal_init_objects
The bCardDataCached field (flags field for what we've already read from
the card) was a member of the P11_SESSION struct, while the data was
stored in the P11_SLOT struct. This is probably a bad idea.

Instead, store the flags field in the P11_SLOT struct as well, so they
get cleaned up together with the card data, rather than when the session
is destroyed.

In addition, switch the check for "has the CDF been read from the card
yet" from a global variable to a per-slot flag in the same slot flags
field.

This makes the sign_state test work again, but not the readdata_sequence
one. That's up next.