Fix GetTokenInfo checks
The PKCS#11 standard says the string here isn't NULL-delimited; it
doesn't say that it can't contain any NULLs.

Since we pass the serial number as binary data, it may contain NULLs,
and in fact does for the card that is in the OS X buildbot machine.

So, remove the check here, so that running the test suite won't fail
anymore.