Turn things around
Rather than having the OCSP code ask OS-specific code for a proxy and use
libcurl to do the HTTP request, have the OCSP code ask OS-specific code
to do the HTTP request instead. This should make things work more easily
where figuring out the correct proxy is harder to do than to just ask
the os to do the damn HTTP request.

Still segfaults in the treeview update, but that's for later.