Fix segfault
Since tst_helper() wants to free() the values that we pass to it, we
must ensure that if we call tst_set() several times, we don't reuse the
pointers that we pass to it.

The documentation of g_main_context_invoke() doesn't say anything about
ordering of functions, so we must assume that ordering is not defined.
As such, copy the value three times and let it be free()'d three times.
Not as efficient as we'd like, but at least it has the advantage of not
segfaulting...