Check that a mech type is actually known to us
There are many mechanism types. If we run C_GetMechanismInfo with a type
that isn't in the table of types that we know about, the for loop
currently assumes the table of elements is zero-terminated.

However, it isn't, so when we call C_MechanismInfo with a garbage type,
we would continue to loop. By pure chance, the table lines up nicely
with the function arguments that the loop ends up matching the type
function argument with itself, but then obviously we return garbage to
the caller.

Fix by explicitly checking whether we're still within the table, and
immediately bailing out if not. While at it, remove the bogus check for
zeroed members, since the table isn't zero-delimited anyway.