Remove undefined behaviour
Assigning a pointer to a table element without first checking whether
the pointer falls within the bounds of that element triggers undefined
behaviour. The clang C++ compiler aggressively optimizes away this loop
because of that, even though we don't actually ever dereference the info
variable in that case.

Fix by turning the check for table bounds and contents around.