Provide a fallback if reading /proc doesn't work
The dialogs call readlink() on /proc/<getppid()>/exe to figure out which
process is trying to read the card.

This works, but:
- It is Linux-specific
- It doesn't work if /proc isn't mounted for some reason.

Currently, the dialog fails when either of those conditions isn't met.
Rather than doing so, it makes more sense to just tell the user what we
do know, and produce a (possibly somewhat awkward) message that doesn't
have all the information we would've wanted to have.