Make error messages when some part of the verification fails clearer
"unauthorized" basically means "you did something wrong". It could be
that the certificate is not the right one, or that an algorithm used to
send the certificate has was not the correct one, or a number of other
things. Just showing "unauthorized" in that case is not very useful.

Also, fix the grammar of the "try again later" error message.