Fix compiler warnings
In most cases, this is just a matter of explicitly casting things
around, but in one case we had to fix a case of use-after-free. The
proper fix required a change of function signature in the Unix version
of the OCSP code, so this means the Linux builds will fail. Will fix
that soon.