Hook up new function so it gets called
Change the definition of find_first_slot() so it takes an extra parameter: the number of slots that were found. This is only guaranteed to be valid if with_token was false, otherwise it is the number of slots with a token that we looked at.

If the number of slots is not the same as previously, call readers_changed so the frontend can update its menu.