Fix compiler warning
CK_UTF8CHAR is defined as "unsigned char" for some unfathomable reason, which makes the compiler moan about sign reversal.

The C standard doesn't actually define whether char is signed for a very good reason, so tell cryptoki to go **** itself and just define things as char*. We do a memcpy() on this anyway.