Don't use default visibility
In gcc lingo, "default" visibility means to export every symbol from a
shared object. We don't want this. Instead, use "hidden" visibility by
default, and explicitly mark all functions that we want to use from the
UI with default visibility.

While at it, ensure that all such functions are nicely
namespace-prefixed, and that they're in the correct header file.

This leaves just one function, but that's one which needs to be cleaned
up by our Windows colleague, so hey.