Use an alternate state machine diagram
We have a state machine diagram which explains for us how the state
machine works; but it might be useful for API users if they were told
which function has which effect on the state machine.

Create a copy of the original state machine diagram but change the event
trigger labels so they refer to functions.