Fix ordering
eidfilepreview requires libxml, which therefore needs to be built before we build it. However, building libxml was only done as part of the build of the eID Viewer, which meant that from a clean build, eidfilepreview couldn't be built.

Fix by making the "build openssl and libxml2" step a part of the build of eidfilefilepreview, rather than the eID Viewer one.