Don't return CKR_TOKEN_NOT_RECOGNIZED from C_GetSlotList()
GetSlotList() is supposed to return a list of slots that have a token.
Sometimes, one of these tokens may not be one of our tokens. We don't
need to tell the caller of GetSlotList this -- they don't care (at least
not yet).

So, filter out use of CKR_TOKEN_NOT_RECOGNIZED.