Make the description an array again
The description member of struct _slotdesc was originally an array. When
converting stuff there so it would use the EID_CHAR type rather than the
char one, Frederik made it an extra allocation; but this created an
issue with the GTK+ viewer where the array of slotdesc structs would be
copied for another thread, but the descriptions wouldn't be.

Fix by using the new macros created in the previous buffer.