Fix return value confusion
When C_GetSlotList receives a NULL pointer as its second parameter, it
will never return CKR_BUFFER_TOO_SMALL (presumably because there is no
buffer to be too small). As such, the body of code that assigned the
first slot through autodetection was never called, and so we couldn't
find a slot unless we manually selected one.

Fix.