Ship the FF47 version, too
Firefox 47+ has changed some of the internal APIs which make it a
requirement that we update our add-on. However, the new (signed) add-on
is marked as "for firefox 47 and above", so doesn't work with older
versions of firefox (such as the current Extended Support Release (ESR)
which Debian, at least, still ships).

So ship them both. Hopefully firefox is smart enough to pick the right
one... we'll know soon enough :-)