Run indent over our code
This touches a *lot* of files, but there is no functional change, only a
code style change.

There is no guarantee that we will ever run indent again; however,
before this commit the code was wildly inconsistent in formatting, which
makes things harder to read than necessary. With this, we'll at least
be consistent.