Don't overload <conversion.h>
The SLEEP macro is a macro, and has nothing to do with string
conversions, for which conversions.h exists.

Since conversions.h is internal to the library, and the SLEEP macro may
be interesting outside the library, move it to the <eid-viewer/macros.h>
file, so that it can be used safely.

To make sure that things continue to compile, add an include directive
for p11.c.