Change return value of p11_lock()
Previously, p11_lock() would sometimes return
CKR_CRYPTOKI_NOT_INITIALIZED. However, this is no longer the case, and
as a result, p11_lock() can now only return CKR_OK.

Coverity complains about us not checking the return value of this
function in a few locations. While it would be reasonable to add the
return value check in those locations, it's cleaner to just drop the
return value (and the dead code in the other locations), since this
function currently cannot fail anyway.

While at it, fix one incorrect usage of this function (where it
specifies a return value that isn't in the declaration) as well as a
minor spelling error.

Coverity CID#18483, and probably others.