Special-case firstnames
Since the "first names" are split over two fields, that means we need to concatenate them. While we can simulate that using two labels for the viewer's main window, this is unfortunately not possible for the PrintView.

To work around that, special-case the firstnames field in the [PrintOperation runOperation] method, so that at least things are aligned properly.

As a side effect, this also does away with the slightly awkward alignment that we do see in the main window.