Remove all references to the "Save as <type>" options
The previous version of the viewer had the option to save the .eID data
as either a .csv or a .eid file. The current viewer only supports the
latter.

Initially, we wrote our UI so that the option to select a .csv file
would still be available. Since we don't actually provide that option
anymore, however, remove it from the user interface.