Don't fail on PIN change operations
The current implementation of the "change PIN" operation does not do a
C_Login first. As a result, the C_Logout here will fail, which causes a
state machine transition failure and hence an error message for the
user, even though the actual pin code change operation was successful.

Still do the logout (so that the "pin test" doesn't keep us logged in
when that isn't necessary), but ignore any error codes that may occur.