Refactor, and add possibility to ask about more props
We may want to hide certain fields in the UI depending on whether they
exist on the type of card we're using.

So, refactor the is_string() function to have it use an array, and add
more fields to that array.

Not entirely pretty, because we run through loads and loads of string
comparisons this way. Hmpf. Will probably refactor this another time at
some undefined point in the future.