Don't create objects for data that doesn't exist on the card.
Previously, if a tag did not exist in either the address or the identity file on
the card, we would still create the object in the PKCS#11 API; querying that
object's value would then return a length of zero, and no data.

This worked fine until the boolean field with tag 0x14 in the identity file was
created (the "member of family" field); this field is either present on the card
(with length zero and no value) to signal "true", or it is absent (i.e., the tag is
not used in the file) to signal "false".

Given the above, before this commit there is no discernible difference for a
user of the PKCS#11 API between either state of this field.

With this change, we change the behaviour of the PKCS#11 to not create
objects for fields that do not exist on the card. The result will be that a
FindObjects operation for fields not present will not return any data, thereby
making it possible to see a difference between the "true" and "false" states.