Allow searching without side effects
Previously, searching for an NSTextView had the side effect that the
NSTextView would be assumed to contain data, and so would be cleared
when the card is removed. This is not necessarily always the case.

Add a BOOL to allow for not doing so, when needed, and only mark
NSTextView fields as "needs to be cleared" from the newstringdata
method.