Update constraints when showing foreigner fields
When inserting a foreigner card, the total number of fields on the
screen is larger than when using an eID card.

Since we want to make sure that things still look good and are nicely
centered, update constraints on the vertical line, so that the new
fields are part of the new "center" location.