Avoid NULL-pointer dereference
Apparently, in an init function, the "self" variable may be nil for some
strange reason. Since assigning to a member variable using an assignment
to _ov tries to dereference self, this may result in a NULL pointer
dereference.

Convert to using the setOv message instead, which can safely be sent to
nil.