Be more future-proof
Use a function to get at a version number rather than a direct string.
This should allow a future version of the add-on to provide the version
number through content scripts and similar things.

It's not worth it to do this right now; but perhaps at some undefined
point in the future, we might also be able to provide other
functionality with this.