Remove dead code
CReader had two member functions "SetEventCallback" and
"StopEventCallback" which were meant to run things upon insertion or
removal of a card from a slot, but this is never called.

The CThreadPool class was used only in support of these two functions.

Remove all of that.

Since this removes a few includes that indirectly included thread.h, add
the now-missing #include statement to the files that need it