GetType(): improve API
GetType() was an abstract virtual function in CCard which returned the
card type in a one-liner in every class. However, that information was
also available in the m_cardType member variable in CCard.

Rework GetType() to be a non-virtual function that just returns the
m_cardType variable, and remove all the virtual implementations.

While at it, move this implementation to the header file of CCard, so
that the compiler can optimize away the function call.