Don't dereference NULL pointers
getenv("HOME") may return NULL if we didn't log in properly somehow.
This is highly unlikely, but not impossible.

As such, check for that, and use getpwuid_r() to figure out our home
directory if we don't have a HOME environment variable.

If that fails too, throw an exception.

Coverity CID#173125