Unbreak eid-viewer test suite
The "pinop" test made a few assumptions about how the backend worked.
These are no longer valid, since we moved the responsibility of calling
the mainloop function from the UI to the backend. Therefore, the
assertion is now no longer valid, and in fact fails in our CI system.

Remove the offending bit, as well as the eid_vwr_poll() call (which does
not cause problems, but is not necessary any more)