Update the install name of the PKCS#11 .dylib
If we want to use a .dylib from our own bundle, we need to tell OSX to go looking in the right spot, otherwise it will always look in /usr/local/lib rather than in our own bundle.

While at it, refresh libbeidpkcs11 version number to 4.2.1 rather than 4.1.19.

TODO: make it an actual version number instead of having it be part of the name. Some other time.