Install XSD in viewer, not framework
Unfortunately, the NSBundle stuff only looks at the executable, not at the framework that was loaded. As such, the relevant resource needs to be installed in the app bundle, not in the framework bundle.

This means that anyone wanting to use the BeidView framework needs to also ship the XSD if they want to open files. That's not a critical fault, but it is something we need to document.