Minidriver: Remove support for deprecated hash algs
Our various middlewares still support officially deprecated algorithms,
including but not limited to:

- md2 (minidriver only)
- md4 (minidriver only)
- md5
- sha1

To allow use of such unsafe hashes is asking for problems. As such, we
should remove them.

This commit removes support for the MD2, MD4, and MD5 algorithms from
the minidriver, as well as the SHA1 one.