Explicitly use the Release directory
Unfortunately, there's a bug in xcode which means it won't automatically
update the source location from Debug to Release or vice versa when the
target configuration is updated.

Do the ugly thing (in preparation for a real fix that may or may not
happen later) and use the Release directory explicitly for now.