Revert "Park, do not eject"
This reverts commit 803fc009d2b7b903357e763c40ad06094a75cd78.

The time it takes to completely remove the card from the reader ensures
that we always see it as "removed" by the time the function returns, if
we tell the robot to eject rather than just park.

It's a bit faster, but that doesn't buy us much if things break too
often in that case.