Avoid a g++-7 warning
See
https://developers.redhat.com/blog/2017/03/10/wimplicit-fallthrough-in-gcc-7/
for more detail.

Add comments to avoid the warning. In one case, we already had a
comment, but it wasn't being picked up by the parser. Reword that one.

In another case, the comments were absent, since the code was pretty
obvious to any human being ("if we don't need *all* data, stop
processing here"). However, it wasn't obvious to g++, so add that
comment to make it so.

Thought: C++17 adds "[[fallthrough]];" as a standardized version of the
above. Maybe add that at some point?