Fix memset call
The "buffer" variable is

wchar_t buffer[MAX_BUFFER_LEN]

where sizeof(wchar_t) is unlikely to be 1. Doing a memset on that buffer
to the number of elements (rather than the number of bytes) therefore is
wrong.

Multiply MAX_BUFFER_LEN by sizeof(wchar_t) to remedy.

Caught by gcc-7 and -Wmemset-elt-size