Attach the sheet to the window rather than the app
Attaching a sheet to an app causes it to be marked as a problem message,
which causes a "warning" sound to be played. This is not what we want.

Instead, attach it to the window, rather than the app. This makes it
still modal, but does not cause the sound anymore.

As a minor side effect, the sheet is shown twice, which means that when
it may disappear, it pops up again (since we queued it twice, rather
than just one time). While we could fix this by checking whether the
sheet is already being shown before asking for it to be shown again,
that seems like overkill, so we don't do it.

This closes Fedict/eid-mw#33.