Mark this as requiring Firefox 57
It actually requires 58 to do anything really useful, but in 57 at least
we can say to a webpage that we can't do much.

The older legacy extension, however, *can* do something useful on older
versions of Firefox, so mark those as not supported by this one, so that
at least we don't make the fuckup worse than it already is.