Update OCSP hashing signature
Earlier, the OCSP server did not (yet) support SHA256. It does now, so
update the code so that it checks with SHA256 rather than SHA1.

While at it, drop the whole check for supported algorithms. The OCSP RFC
says that the server should support at least the algorithms used for the
certificates themselves, but we already know it supports SHA256, and
there's really no reason to fall back on SHA1.