Merge "eID Viewer.xcodeproj" into beidmw.xcodeproj
If there was ever a good reason for keeping the two projects separate,
I've forgotten about it.

Meanwhile, it's causing real issues with the viewer not being able to be
properly debugged etc, and it including an older version of the
middleware rather than the one we actually need.

As such, merge the eID Viewer targets into the beidmw.xcodeproj project,
and drop the eID Viewer.xcodeproj project.