Fix parameter type
C_GetAttributeValue takes a CK_OBJECT_HANDLE, not a
CK_OBJECT_HANDLE_PTR.

While this test just checks for correct reply in case we send things in
the wrong order (so this error should not cause false positives or false
negatives), sending invalid data is not proper and we should not do it.

Fix.