Include headers with declarations
We had a header file containing some declarations of functions
implemented in this file, but didn't include it here.

This may cause problems if we ever choose to update the declaration but
forget to update the header file to match, or vice versa. Include them
to make sure that that never happens.