Remove "Obsoletes:" header
We added it there originally so as to avoid issues when upgrading from
the Java viewer to the native one, but it confuses DNF into believing
that the 32-bit version of the viewer obsoletes the 64-bit one (or vice
versa).

Since the demise of the Java viewer is a few years old now anyway, drop
the Obsoletes header from the spec file to fix.