Don't sleep unless we need to lock
The eID middleware tries to open a file, but will loop on doing so if it
receives EACCES in the assumption that the EACCES was due to another
process locking the file.

In the presence of mandatory access control frameworks such as AppArmor
and SELinux, this may not be the case, in which case we'd be stuck in an
infinite loop.

Change things so we only loop if we were trying to lock the file in the
first place.