Automatically set the version number when necessary
If a field is being used that's only supported by a version 4.1 .eid
file, we want to automatically set that version number in the XML
output.

Do so by checking in the caching layer when we set a field that requires
a particular version; if that happens, we set the version number.

Write it in a future-compatible way so that if we ever need to add other
fields, we can do so.