Don't produce an error on the xml label
We write the "xml" label to the cache, too. Since we now check every
cache element on whether it has a required version, that produces a
spurious error message "unknown label".

Need to find a proper way to deal with that, but for now we'll just
special-case that label here.