Stop providing our own versions of library functions
For some reason, we carry our private vasprintf() implementation. When
trying to compile with clang 6 or above an -Wall -Werror, we tripped
over a warning at configure stage that caused vasprintf() to not be
detected, which resulted in this being picked up -- but also in clang
complaining about the (private) vasprintf declaration not matching the
(system) vasprintf one.

It makes no sense to provide a private vasprintf() implementation
anymore; we don't support any OS anymore that doesn't provide it. As
such, drop the private implementation, and tweak the configure script so
that failure to find one of the required functions is a hard error,
rather than a "oh let's try to work around it" kind of thing.