Don't abort writing, just skip an element
When we don't have the required data for a particular element, we should
skip it. However, the current implementation would abort writing the
parent element, rather than just skipping the element we didn't need.

This doesn't currently result in invalid XML files since we don't
actually have any element that might need to be skipped which doesn't
have other non-skipped data following it, but that might be the case in
the future. Better do it right now.