Don't check for ftime()
ftime(3) says:

  This function is obsolete. Don't use it.

Also, FreeBSD doesn't have it.

Apparently we already patched it out, because a simple grep over the
code reveals we're not using it anywhere. However, configure still had
it.

Since failure to find a function is now a hard error, that means that
we're now failing on FreeBSD for no particularly good reason.

Drop the check for that function to fix.