Improve optional things
We dropped optional parent elements if non-optional child elements were
not present. However, we ignored the presence (or absense) of optional
child elements, which seems to make no sense.

Instead, change it so that we show an optional parent element if any
child element has data, optional or otherwise. If we miss a non-optional
element, we'll still get an error message as usual.