Update package dependencies
Now that Firefox ESR 60 has been released, most distributions have
dropped ESR 52 for which the older XPC-based add-on was required.

The sole exception to this rule is RHEL6. One might wonder whether it's
worth keeping a package for a distribution that is that ancient; but
since it costs us next to nogthing, might as well keep the XPC extension
there.

Everywhere else, drop it and switch to the webextension.