Don't fail on Fedora 29+
We configure our PKCS#11 library into the p11-kit configuration, so that
applications (other than Firefox) can easily make use of our BeID
module.

In addition, we ship a Firefox add-on which uses the browser.pkcs11 API
to load our PKCS#11 module into Firefox automatically, and sideload that
extension.

Recent versions of Fedora have configured Firefox to autoload the
p11-kit-proxy.so module so that all registered PKCS#11 modules are
available on the system by default, which is great; but it does have the
unfortunate side effect that if our add-on tries to load the BeID
PKCS#11 module, that fails (because it has been loaded already).

Work around this issue by adding a configuration giving access to the
p11-kit-proxy.so module to our add-on; if (and only if) it fails to load
the BeID PKCS#11 module, we can then check if the p11-kit-proxy module
is loaded, and if so, assume that that was the reason and move on
without loading the module.

(cherry picked from commit db57066af9954be941a0d402cc3cd088d5816cd4)