Also update the getModuleSlots stuff
Now that the module may not be accessed directly, we may need to
redirect module stuff to the p11-kit module.

Since our add-on only cares about Belgian eID cards, ensure that if
another module is loaded through p11-kit, we filter those out by
skipping any slots that have a token but don't have "Belgium Government"
as the manufacturer.