Update state change handler
After the changes in commit c8413579cc57d3c9fbc04d0eabd40e66e75e6e7d,
the GTK viewer didn't enable the various card-related buttons anymore.
This was due to the fact that we disable those buttons unless the final
state was TOKEN_WAIT, in the assumption that there was no child state of
that state.

Since this actually is the case now, change the state handling so that
it does the necessary changes in the child state rather than the parent
one.

TODO: validate that the same problem doesn't exist on macOS