Don't store a nil value
When the resources are unavailable (which should not happen, but...) then NSImage's initByReferencingFile: method will return nil, which causes a crash later on.

Fix by creating an empty 10x10 image in that case.

(of course, we should also not fail to include the resources, but this doesn't hurt)

(cherry picked from commit dd4f6a4bfbf33bf691da9c55e11b456343df60e1)