Do fuzzing with libFuzzer
In order to be able to fuzz, we need to instrument our libraries, which
requires a few extra compiler flags. Add them, but only if
--enable-test-fuzzer is provided to configure. Also, add a function to
inject the fuzzed data in that case.

libFuzzer also requires a function that injects the random data and then
runs some code. Write it, in a generic way so we can reuse it for other
stuff later.

TODO: try different fuzzers (AFL, perhaps) and different files (not just
the ID file). Later.