Remove P15Correction and friends
CP15Correction and CBeidP15Correction are two classes that allow
overriding of the output of the PKCS#15 parser for specific cards. This
is written in a generic way, so it works for multiple cards.

Since we only support one card, just drop the class and override the
parsed data in the parser directly.

TODO: drop the parser -- we know what the card looks like, and
reading/parsing a PKCS#15 file takes time. Not just yet.

TODO: drop the necessary files from the Xcode and Visual Studio
projects. Up next.