Fix .so visibility
We were not using any -fvisibility parameter on libbeidpkcs11.so, which
meant that the backwards-compatible "default" visibility was in effect;
as a result, all functions (also those that were meant to be internal to
the PKCS#11 module) were being exported. This is not ideal, for multiple
reasons:

- It is bad for performance;
- It might allow internal functions to be accessed when that is not the
  intent
- It makes the ABI unnecessarily broad.

These extra methods are not documented and should (hopefully) not be
used by anyone, so just drop them without notice. If that turns out to
be a problem, then, well, uhm, we'll see what we can do...